Move helper class inside common base class to reduce public space pollution
