+++ server/src/main/java/org/apache/druid/segment/indexing/granularity/ArbitraryGranularitySpec.java
@@ -26,7 +26,6 @@ import com.google.common.collect.PeekingIterator;
 import org.apache.druid.java.util.common.IAE;
 import org.apache.druid.java.util.common.granularity.Granularities;
 import org.apache.druid.java.util.common.granularity.Granularity;
-import org.apache.druid.segment.indexing.LookupIntervalBuckets;
 import org.joda.time.Interval;
 
 import javax.annotation.Nullable;
